+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Updated dependencies: `std`, `deno_dom`, `parcel_css`, `imagick`, `liquid`.
 - Ignore `deno.jsonc` file by default [#209].
 - Ignored files starting with `.` or `_` in directories copied statically (with `site.copy()`).
+- Copy static files starting with `.` or `_` if they are are explicit in the config file [#210].
 
 ## [1.10.0] - 2022-06-16
 ### Added
@@ -1626,6 +1627,7 @@ The first version.
 [#203]: https://github.com/lumeland/lume/issues/203
 [#207]: https://github.com/lumeland/lume/issues/207
 [#209]: https://github.com/lumeland/lume/issues/209
+[#210]: https://github.com/lumeland/lume/issues/210
 
 [1.10.1]: https://github.com/lumeland/lume/compare/v1.10.0...HEAD
 [1.10.0]: https://github.com/lumeland/lume/compare/v1.9.1...v1.10.0